firebase storage upload working
download URL is generated as well. will need to run an API update call to the recipe to add the photo link. i have no idea on authentication, i think anyone can view the pic with the link because of the auth token in the link. i dont know if random users can upload maybe it only lets me do it because im admin. needs testing and validation.